+++ src/sha1/digest.rs
@@ -197,7 +197,6 @@ impl UpperHex for Digest {
 }
 
 #[cfg(feature = "error")]
-#[cfg_attr(docsrs, doc(cfg(feature = "error")))]
 impl TryFrom<&str> for Digest {
     type Error = Error;