+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 
 - Changed crate description.
+- Changed `doc_cfg` feature with `doc_auto_cfg`.
 
 ## [0.3.0] - 2023-06-01